refactor: split command_contract lab routing out of lab type definitions
command_contract/lab.rs mixed two concerns: the Lab/runner contract TYPES that
core imports (LabCommandContract, RunnerWorkload, handoff schemas, ~everything
core needs) and CLI-coupled routing (two impl Commands blocks plus the
extension-id helpers) that depend on cli_surface::Commands and command handler
modules. That mixing forced a commands dependency onto a file core relies on --
part of the command_contract <-> commands cycle blocking a homeboy-cli
extraction.

Move the CLI-coupled routing into a new command_contract/lab_routing.rs:
- both `impl Commands` blocks (lab_route_contract, descriptor resolution)
- the extension_ids helpers that take agent_task/review Args
- apply_lab_contract_to_descriptor (output.rs import repointed)

lab.rs is now free of commands/cli_surface in production (only a doc comment
and test module reference them). A few lab.rs consts/helpers were promoted to
pub(crate) so the relocated routing can reach them across the module boundary.
No behavior change; pure code movement.

This is step 1 of the command_contract decomposition that unblocks homeboy-cli.
cargo check --lib and --lib --tests pass.
