+++ src/command_contract/output.rs
@@ -14,7 +14,7 @@
 use crate::cli_surface::Commands;
 use crate::commands::{adapter, file, logs, report, review, runner, runtime, trace};
 
-use super::lab::apply_lab_contract_to_descriptor;
+use super::lab_routing::apply_lab_contract_to_descriptor;
 use super::spec::CommandSpec;
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
